requirements.txt
openai==1.13.3
requests==2.31.0
streamlit==1.28.2
streamlit-lottie==0.0.5
streamlit-webrtc==0.47.1
streamlit-chat==0.1.1
PyYAML>=6.0.0
doit==0.36.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
dotted-dict==1.1.3
python-docx==1.1.0
pytest==6.2.5
openpyxl==3.1.2
chromadb
pysqlite3-binary
